backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
pillow
python-multipart
opencv-python 
numpy 
pillow 
exifread 
scikit-image 
scipy